redbug_dtop: don't crash on mnesia ext tables
If you try to run redbug:dtop() in a system with mnesia ext tables,
it crashes with a badarith because redbug_dtop:ets_size/1 multiplies
the atom 'undefined' with an integer.

This happens because get_term_storage_type/1 defaults to 'ets' for
not explicitly matched storage types. This is wrong for e.g. ext
LevelDB tables: only 'disc_copies' should be mapped to 'ets'.

Also change the type marker 'remote_only' to 'undefined' to better
match its semantics (anything dtop doesn't know how to handle).